Feat: Upgrade project to C++20 and refactor logger macros
This commit upgrades the project's C++ standard from C++17 to C++20.

This change enables the use of modern C++20 features. It is immediately
leveraged to refactor the logger macros to use `__VA_OPT__`. This
provides a more ergonomic logging API by making the `FMT_STRING()`
wrapper automatic, while remaining fully compliant with the C++ standard
and eliminating the need for non-standard extensions.