fix logger refactoring, test now passed. need another critical review
PLATFORM_* need to be used correctly. Should check if sink switch request is necessary by comparing first, and switch from message should be used as well. should include syslog and eventlog tests for different platforms. FileLock should delete .lock file if it is the owner and release is done properly